// entry/src/main/ets/model/UserProfileModels.ts
export class UserProfileInfo {
  uid: number = 0;
  portrait: string = '';
  userName: string = '';
  nickName: string = '';
  gender: number = 0;
  age: number = 0;
  threadNum: number = 0;
  postNum: number = 0;
  fanNum: number = 0;
  followNum: number = 0;
  isVip: boolean = false;
  isFollowing: boolean = false;
  sign: string = '';
  tiebaUid: number = 0;
  agreeNum: number = 0;
}